changed deposits and borrows to u128

+++ programs/pathfinder/src/instructions/borrow.rs
@@ -123,9 +123,9 @@ impl<'info> Borrow<'info> {
     accrue_interest(market, config)?;
 
     if assets > 0 {
-      shares = to_shares_up(assets, market.total_borrows, market.total_borrow_shares)?;
+      shares = to_shares_up(assets, market.total_borrows as u64, market.total_borrow_shares)?;
     } else {
-      assets = to_assets_down(shares, market.total_borrows, market.total_borrow_shares)?;
+      assets = to_assets_down(shares, market.total_borrows as u64, market.total_borrow_shares)?;
     }
 
     // check if user is solvent after borrowing
@@ -156,7 +156,7 @@ impl<'info> Borrow<'info> {
     // Update market total borrows
     market.total_borrows = market
       .total_borrows
-      .checked_add(assets)
+      .checked_add(assets as u128)
       .ok_or(MarketError::MathOverflow)?;
 
     // transfer tokens to borrower
@@ -191,7 +191,7 @@ pub fn is_solvent(
   let price = oracle_get_price(&market.oracle, &oracle_ai, false)?;
 
   // Calculate borrowed amount by converting borrow shares to assets, rounding up
-  let borrowed = to_assets_up(borrow_shares, market.total_borrows, market.total_borrow_shares)?;
+  let borrowed = to_assets_up(borrow_shares, market.total_borrows as u64, market.total_borrow_shares)?;
 
   // Calculate max borrow amount based on collateral value and LTV factor
   let max_borrow = (collateral_amount as u128)
