+++ programs/pathfinder/src/instructions/interest_rate.rs
@@ -35,7 +35,7 @@ pub fn get_rate(market: &Account<Market>) -> Result<(Decimal, Decimal)> {
 
   // Safe "unchecked" cast because the utilization is smaller than 1 (scaled by WAD).
   let utilization: i128 = if market.total_deposits > 0 {
-    w_div_down(market.total_borrows, market.total_deposits)? as i128
+    w_div_down(market.total_borrows as u64, market.total_deposits as u64)? as i128
   } else {
     0
   };
